+++ spec/kitchen_spec/directions/bake_figure_spec.rb
@@ -144,7 +144,7 @@ RSpec.describe Kitchen::Directions::BakeFigure do
     book_containing(html:
       one_chapter_with_one_page_containing(
         <<~HTML
-          <figure id="someId" class="mechanism-figure">
+          <figure id="someId" class="mechanism-figure scaled-down">
             #{figure_caption}
             #{figure_title}
             <span data-type="media" id="otherId" data-alt="This figure shows pieces of a ...">
@@ -303,7 +303,10 @@ RSpec.describe Kitchen::Directions::BakeFigure do
   end
 
   it 'logs a warning' do
-    expect(Warning).to receive(:warn).with(/warning! exclude unnumbered figures from `BakeFigure` loop/)
+    expect(Warning).to receive(:warn).with(
+      /warning! exclude unnumbered figures from `BakeFigure` loop/,
+      { category: nil }
+    )
     described_class.v1(figure: book_with_unnumbered_splash.figures.first, number: 'blah')
   end